requirements.txt
pycryptodome==3.9.4
cython==0.29.14
stdeb==0.9.0
wheel==0.33.6

#_gid	GA1.2.1892352145.1513594682	.wings.io	/	2017-12-20T10:05:31.000Z	31			
#_ga	GA1.2.1158346106.1511987264	.wings.io	/	2019-12-19T10:05:31.000Z	30			
#_ga	GA1.1.1158346106.1511987264	wings.io	/	2019-12-01T12:19:08.000Z	30			
#__gads	ID=0079b47909cd1f3f:T=1512041675:S=ALNI_MaduDPbymbCOqDUxlsnxgUttwrnlQ	.wings.io	/	2019-11-30T11:34:35.000Z	75			
#__cfduid	d99883e767f1671ae125e4f1a537cb4591511987263	.wings.io	/	2018-11-29T20:27:43.533Z	51	✓		
